Created possibly better heuristic
disregards 2 and 3 in a row, when there is no way to get 4 in a row with
them.

Also discovered a weird bug where the ai would not see 4 in a row.  They
were placed diagonally up to the right from the left one to the bottom
most in the middle